requirements.txt
streamlit
GitPython
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
tree-sitter-python==0.21.0
networkx
google-generativeai
